Fix for toolforge.org migration of other tools
The LDAP tool migrated to toolforge.org, which changed the HTML it
serves from something like

    <li><a href="https://wikitech.wikimedia.org/wiki/User:Awight">User:Awight</a> (<a href="/ldap/user/awight">more info</a>)</li>

to

    <li><a href="https://wikitech.wikimedia.org/wiki/User:Awight">User:Awight</a> (<a href="/user/awight">more info</a>)</li>

i. e. the links within the tool no longer start with /ldap/ because the
tool now has its own subdomain. This broke our analysis of LDAP group
memberships; adjust the regex to match it, and then while we’re at it,
also change the links we emit ourselves to point to toolforge.org.